[apps] Fix calculation buffer overflow
The check in pushEmptyCalculation was wrong because the space wrapped

+++ apps/calculation/calculation_store.h
@@ -46,7 +46,7 @@ class CalculationStore {
   Shared::ExpiringPointer<Calculation> calculationAtIndex(int index) const;
   Poincare::Expression ansExpression(Poincare::Context *context) const;
   size_t bufferSize() const { return m_bufferSize; }
-  size_t remainingBufferSize() const {
+  int remainingBufferSize() const {
     return spaceForNewCalculations(endOfCalculations()) + sizeof(Calculation *);
   }
 
@@ -82,7 +82,7 @@ class CalculationStore {
   }
   char *endOfCalculationAtIndex(int index) const;
   /* Account for the size of an additional pointer at the end of the buffer. */
-  size_t spaceForNewCalculations(char *currentEndOfCalculations) const {
+  int spaceForNewCalculations(char *currentEndOfCalculations) const {
     return (pointerArea() - currentEndOfCalculations) - sizeof(Calculation *);
   }
